Moved temperature fetches into TemperatureService.js, fixed problems with undefined temp.
